+++ Snake/Program.cs
@@ -29,7 +29,7 @@ class Program
         world.Create_Food();
         world.Create_Food();
 
-        AISnake aISnake = new AISnake(world, new Position { x= WorldWidth-4, y = WorldHeight-4});   
+        TheGreatAI aISnake = new TheGreatAI(world, new Position { x= WorldWidth-4, y = WorldHeight-4});   
         world.gameObjects.Add(aISnake); 
 
 
@@ -103,6 +103,7 @@ class Program
 
     static void Main(string[] args)
     {
+     
         // Vi kan ev. ha någon meny här, men annars börjar vi bara spelet direkt
         Loop();
     }
